Rename habits.js to Habits.js to keep with convention
