+++ api/controllers/Crud.js
@@ -1,29 +1,25 @@
-const Habits = require("../models/habits")
-
+const Habits = require("../models/Habits");
 
 //list habits
 
-async function showHabits(req,res){
-    try{
-        const userHabits2 = await Habits.findById(req.params.UserId);
-        //const userHabits2 = await Habits.findById;
-        res.status(200).json(userHabits2)
-
-    }
-    catch (err){
-        res.status(404).send({err})
-    }
+async function showHabits(req, res) {
+  try {
+    const userHabits2 = await Habits.findById(req.params.UserId);
+    //const userHabits2 = await Habits.findById;
+    res.status(200).json(userHabits2);
+  } catch (err) {
+    res.status(404).send({ err });
+  }
 }
 
 // list specific habits
-async function specificHabit(req,res) {
-    try{
-        const a = await Habits.specificHabits(req.params.UserId,req.params.habit);
-        res.status(200).json(a)
-    }
-    catch(err){
-        res.status(404).send(err)
-    }
+async function specificHabit(req, res) {
+  try {
+    const a = await Habits.specificHabits(req.params.UserId, req.params.habit);
+    res.status(200).json(a);
+  } catch (err) {
+    res.status(404).send(err);
+  }
 }
 
 // async function showOne(req, res) {
@@ -37,50 +33,45 @@ async function specificHabit(req,res) {
 
 //create a habit
 
-async function createHabit(req,res){
-    try{
-        /*const newData = {
+async function createHabit(req, res) {
+  try {
+    /*const newData = {
             UserId:req.params.UserId,
             ...req.body
         }
         const newData2 = await Habits.createHabit(newData);
         res.status(201).json(newData2);*/
-        const newResult = await Habits.createHabit(req.body)
-        res.status(201).json(newResult)
-        
-             
-        } catch(err){
-            res.status(404).send(err);
-
-        }
-    }
-
+    const newResult = await Habits.createHabit(req.body);
+    res.status(201).json(newResult);
+  } catch (err) {
+    res.status(404).send(err);
+  }
+}
 
 //update habits
 async function update(req, res) {
-    try {
-        const habit = await Habits.specificHabit(req.params.id)
-        const updatedStreak = await habit.update()
-        res.json({streak: updatedStreak})
-    } catch (err) {
-        res.status(500).send({err})
-    }
+  try {
+    const habit = await Habits.specificHabit(req.params.id);
+    const updatedStreak = await habit.update();
+    res.json({ streak: updatedStreak });
+  } catch (err) {
+    res.status(500).send({ err });
+  }
 }
 
 //delete habits
 async function destroy(req, res) {
-     try {
-        // const habit = await Habits.findById(req.params.id);
-        await Habits.delete(req.body);
-        res.status(204).json('Habit was deleted')
-    } catch (err) {
-        res.status(500).json({err})
-    }
-    
+  try {
+    // const habit = await Habits.findById(req.params.id);
+    await Habits.delete(req.body);
+    res.status(204).json("Habit was deleted");
+  } catch (err) {
+    res.status(500).json({ err });
+  }
 }
 
 // async function destroyAll (req, res) {
-  
+
 //     try {
 //        const habits = await Habits.all;
 //        await habits.destroy();
@@ -88,7 +79,7 @@ async function destroy(req, res) {
 //    } catch (err) {
 //        res.status(500).json({err})
 //    }
-   
+
 // }
 
-module.exports = { destroy, showHabits, update, specificHabit, createHabit}
+module.exports = { destroy, showHabits, update, specificHabit, createHabit };
